requirements.txt
numpy
tensorflow
opencv-python
pillow